requirements.txt
asgiref==3.5.1
Django==4.0.4
python-dotenv==0.20.0
sqlparse==0.4.2
tzdata==2022.1
djangorestframework~=3.13.1